move task to concurrent
