+++ src/core/concurrent/task_engine.h
@@ -35,7 +35,7 @@
 
 #pragma once
 
-#include "service_engine.h"
+#include "core/core/service_engine.h"
 #include <dsn/tool-api/task_queue.h>
 #include <dsn/tool-api/admission_controller.h>
 #include <dsn/perf_counter/perf_counter.h>
