+++ src/core/rpc/rpc_message.cpp
@@ -41,7 +41,7 @@
 #include <dsn/tool-api/message_parser.h>
 #include <cctype>
 
-#include "core/core/task_engine.h"
+#include "core/concurrent/task_engine.h"
 
 using namespace dsn::utils;
