+++ src/core/tests/task_engine.cpp
@@ -33,7 +33,7 @@
  *     xxxx-xx-xx, author, fix bug about xxx
  */
 
-#include "../core/task_engine.h"
+#include "core/concurrent/task_engine.h"
 #include "test_utils.h"
 #include <dsn/tool_api.h>
 #include <gtest/gtest.h>
